# speed up analyze-outcomes stage