--- a/vars/analysis.groovy
+++ b/vars/analysis.groovy
@@ -246,10 +246,10 @@ def process_outcomes() {
     // as test description or test suite was "FAIL".
     if (gen_jobs.failed_builds) {
         sh '''\
-grep ';FAIL;' outcomes.csv >"failures.csv"
+LC_ALL=C grep ';FAIL;' outcomes.csv >"failures.csv"
 # Compress the failure list if it is large (for some value of large)
 if [ "$(wc -c <failures.csv)" -gt 99999 ]; then
-    LC_ALL=C xz -0 -T8 failures.csv
+    xz -0 -T8 failures.csv
 fi
 '''
     }