@daverodgman 2024-01-03
@gilles-peskine-arm I have updated to use `-T0`, please review