@daverodgman 2024-01-04
Test run shows xz taking 15s https://jenkins-mbedtls.oss.arm.com/blue/rest/organizations/jenkins/pipelines/mbed-tls-pr-head/branches/PR-8530-head/runs/12/nodes/8129/log/?start=0